user_interface/monitoring_project/requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
asyncio==3.4.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-crispy-forms==2.1
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
numpy==1.26.4
pandas==2.2.2
prompt_toolkit==3.0.46
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
